Enable cross-representation expression transform for all values
Implement abstract_value_objectt::expression_transform, and prevent
further overriding. Within that, dispatch according to the types of the
operands -

* any are value sets, iterate over all of the combinations and
  dispatch for each of these, producing a value set to contain
  all of the results
* mixed constant and interval, convert all constants to intervals
  and handle as all intervals
* all interval, do interval arithmetic
* all constant, substitute and simplify
