+++ src/analyses/variable-sensitivity/interval_abstract_value.cpp
@@ -282,151 +282,6 @@ exprt interval_abstract_valuet::to_constant() const
 #endif
 }
 
-abstract_object_pointert interval_abstract_valuet::expression_transform(
-  const exprt &expr,
-  const std::vector<abstract_object_pointert> &operands,
-  const abstract_environmentt &environment,
-  const namespacet &ns) const
-{
-  std::size_t num_operands = expr.operands().size();
-  PRECONDITION(operands.size() == num_operands);
-
-  std::vector<sharing_ptrt<interval_abstract_valuet>> interval_operands;
-
-  for(const auto &op : operands)
-  {
-    auto iav = std::dynamic_pointer_cast<const interval_abstract_valuet>(op);
-    if(!iav)
-    {
-      // The operand isn't an interval - if it's an integral constant we can
-      // convert it into an interval.
-
-      if(constant_interval_exprt::is_int(op->type()))
-      {
-        const auto op_as_constant = op->to_constant();
-        if(op_as_constant.is_nil())
-        {
-          auto top_object =
-            environment.abstract_object_factory(expr.type(), ns, true, false);
-          auto top_context_object =
-            std::dynamic_pointer_cast<const context_abstract_objectt>(
-              top_object);
-          CHECK_RETURN(top_context_object);
-          return top_context_object->get_child();
-        }
-        const auto ivop =
-          environment.abstract_object_factory(op->type(), op_as_constant, ns);
-        const auto ivop_context =
-          std::dynamic_pointer_cast<const context_abstract_objectt>(ivop);
-        if(ivop_context)
-        {
-          iav = std::dynamic_pointer_cast<const interval_abstract_valuet>(
-            ivop_context->get_child());
-        }
-        else
-          iav = std::dynamic_pointer_cast<const interval_abstract_valuet>(ivop);
-      }
-      CHECK_RETURN(
-        !std::dynamic_pointer_cast<const context_abstract_objectt>(iav));
-
-      if(!iav)
-      {
-        // If we could not convert the operand into an interval,
-        // e.g. if its type is not something we can represent as an interval,
-        // try dispatching the expression_transform under that type instead.
-        return op->expression_transform(expr, operands, environment, ns);
-      }
-    }
-
-    INVARIANT(iav, "Should be an interval abstract value");
-    interval_operands.push_back(iav);
-  }
-
-  if(num_operands == 0)
-    return environment.abstract_object_factory(expr.type(), ns, true, false);
-
-  if(expr.id() == ID_if)
-    return evaluate_conditional(expr, interval_operands, environment, ns);
-
-  if(num_operands == 1)
-    return evaluate_unary_expr(expr, interval_operands, environment, ns);
-
-  constant_interval_exprt result = interval_operands[0]->interval;
-
-  for(size_t opIndex = 1; opIndex != interval_operands.size(); ++opIndex)
-  {
-    auto &interval_next = interval_operands[opIndex]->interval;
-    result = result.eval(expr.id(), interval_next);
-  }
-
-  INVARIANT(
-    result.type() == expr.type(),
-    "Type of result interval should match expression type");
-  return environment.abstract_object_factory(expr.type(), result, ns);
-}
-
-abstract_object_pointert interval_abstract_valuet::evaluate_conditional(
-  const exprt &expr,
-  const std::vector<interval_abstract_value_pointert> &interval_operands,
-  const abstract_environmentt &env,
-  const namespacet &ns)
-{
-  auto const &condition_interval = interval_operands[0]->interval;
-  auto const &true_interval = interval_operands[1]->interval;
-  auto const &false_interval = interval_operands[2]->interval;
-
-  auto condition_result = condition_interval.is_definitely_true();
-
-  if(condition_result.is_unknown())
-  {
-    // Value of the condition is both true and false, so
-    // combine the intervals of both the true and false expressions
-    return env.abstract_object_factory(
-      expr.type(),
-      constant_interval_exprt(
-        constant_interval_exprt::get_min(
-          true_interval.get_lower(), false_interval.get_lower()),
-        constant_interval_exprt::get_max(
-          true_interval.get_upper(), false_interval.get_upper())),
-      ns);
-  }
-
-  return condition_result.is_true()
-           ? env.abstract_object_factory(
-               true_interval.type(), true_interval, ns)
-           : env.abstract_object_factory(
-               false_interval.type(), false_interval, ns);
-}
-
-abstract_object_pointert interval_abstract_valuet::evaluate_unary_expr(
-  const exprt &expr,
-  const std::vector<interval_abstract_value_pointert> &interval_operands,
-  const abstract_environmentt &environment,
-  const namespacet &ns)
-{
-  const constant_interval_exprt &operand_expr = interval_operands[0]->interval;
-
-  if(expr.id() == ID_typecast)
-  {
-    const typecast_exprt &tce = to_typecast_expr(expr);
-
-    const constant_interval_exprt &new_interval =
-      operand_expr.typecast(tce.type());
-
-    INVARIANT(
-      new_interval.type() == expr.type(),
-      "Type of result interval should match expression type");
-
-    return environment.abstract_object_factory(tce.type(), new_interval, ns);
-  }
-
-  const constant_interval_exprt &interval_result = operand_expr.eval(expr.id());
-  INVARIANT(
-    interval_result.type() == expr.type(),
-    "Type of result interval should match expression type");
-  return environment.abstract_object_factory(expr.type(), interval_result, ns);
-}
-
 void interval_abstract_valuet::output(
   std::ostream &out,
   const ai_baset &ai,
